+++ app/src/main/java/woowacourse/movie/ui/bookinghistory/BookingHistoryDBAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.main
+package woowacourse.movie.ui.bookinghistory
 
 import android.content.ContentValues
 import android.database.Cursor
