+++ app/src/main/java/woowacourse/movie/ui/home/HomeContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.main
+package woowacourse.movie.ui.home
 
 import woowacourse.movie.model.main.MainData
