+++ app/src/main/java/woowacourse/movie/ui/setting/SettingContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.main
+package woowacourse.movie.ui.setting
 
 interface SettingContract {
