+++ app/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
         android:id="@+id/fragmentContainerView"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        android:name="woowacourse.movie.ui.main.HomeFragment"
+        android:name="woowacourse.movie.ui.home.HomeFragment"
         app:layout_constraintBottom_toTopOf="@id/main_bottom_navi"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
