+++ app/src/main/res/layout/fragment_setting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.SettingFragment">
+    tools:context=".ui.setting.SettingFragment">
 
     <TextView
         android:id="@+id/textSettingPushAlarm"
